+++ packages/immutable-arraybuffer/index.js
@@ -1,8 +1,10 @@
 /* global globalThis */
 
-const { setPrototypeOf, getOwnPropertyDescriptors } = Object;
+const { setPrototypeOf, getOwnPropertyDescriptors, defineProperties } = Object;
 const { apply } = Reflect;
 const { prototype: arrayBufferPrototype } = ArrayBuffer;
+// Capture structuredClone before it could be scuttled.
+const { structuredClone: originalStructuredCloneMaybe } = globalThis;
 
 const {
   slice,
@@ -38,12 +40,14 @@ let arrayBufferTransfer;
 
 if (transfer) {
   arrayBufferTransfer = arrayBuffer => apply(transfer, arrayBuffer, []);
-} else if (globalThis.structuredClone) {
+} else if (originalStructuredCloneMaybe) {
   arrayBufferTransfer = arrayBuffer => {
     // Hopefully, a zero-length slice is cheap, but still enforces that
     // `arrayBuffer` is a genuine `ArrayBuffer` exotic object.
     arrayBufferSlice(arrayBuffer, 0, 0);
-    return globalThis.structuredClone(arrayBuffer, { transfer: [arrayBuffer] });
+    return originalStructuredCloneMaybe(arrayBuffer, {
+      transfer: [arrayBuffer],
+    });
   };
 } else {
   // Indeed, Node <= 16 has neither.
@@ -141,6 +145,16 @@ const {
 } = getOwnPropertyDescriptors(immutableArrayBufferPrototype);
 
 setPrototypeOf(immutableArrayBufferPrototype, arrayBufferPrototype);
+
+// See https://github.com/endojs/endo/tree/master/packages/immutable-arraybuffer#purposeful-violation
+defineProperties(immutableArrayBufferPrototype, {
+  [Symbol.toStringTag]: {
+    value: 'ImmutableArrayBuffer',
+    writable: false,
+    enumerable: false,
+    configurable: true,
+  },
+});
 
 /**
  * Transfer the contents to a new Immutable ArrayBuffer