+++ packages/ses/test/error/tame-console.test.js
@@ -6,10 +6,10 @@ const originalConsole = console;
 
 lockdown({ errorTaming: 'safe' });
 
-const { details: X, quote: q, note: annotateError } = assert;
+const { details: X, quote: q, note: annotateError, Fail } = assert;
 
 test('ava message disclosure default', t => {
-  t.throws(() => assert.fail(X`a secret ${666} and a public ${q(777)}`), {
+  t.throws(() => Fail`a secret ${666} and a public ${q(777)}`, {
     message: /a secret \(a number\) and a public 777/,
   });
 });
@@ -58,7 +58,7 @@ test('assert - safe', t => {
   try {
     const obj = {};
     const fooErr = SyntaxError('foo');
-    assert.fail(X`caused by ${fooErr},${obj}`);
+    Fail`caused by ${fooErr},${obj}`;
   } catch (barErr) {
     console.error('bar happens', barErr);
   }
@@ -73,7 +73,7 @@ test('assert - unlogged safe', t => {
   t.throws(() => {
     const obj = {};
     const fooErr = SyntaxError('foo');
-    assert.fail(X`caused by ${fooErr},${obj}`);
+    Fail`caused by ${fooErr},${obj}`;
   });
 });